feat: move packages index page to its own folder and handle displaying packages
